socket_server/requirement.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
idna==2.10
idna-ssl==1.1.0
multidict==4.7.6
PyJWT==1.7.1
python-engineio==3.13.1
python-socketio==4.6.0
requests==2.24.0
shortio==0.1.0
sio==0.1.0
six==1.15.0
typing-extensions==3.7.4.2
urllib3==1.25.10
yarl==1.5.1
